requirements.txt
openai==1.30.5
pydub==0.25.1
python-dotenv==1.0.1
sentence-transformers==3.0.0
soundfile==0.12.1
streamlit==1.35.0
streamlit-audiorec==0.1.3
langchain==0.2.1
langchain-community==0.2.1
langchain-core==0.2.3
langchain_openai
